+++ src/main/index.ts
@@ -5,9 +5,9 @@ import { app, BrowserWindow, dialog, ipcMain } from 'electron';
 import log from 'electron-log';
 import MarkdownIt from 'markdown-it';
 
-import { Mediator } from '../common/Mediator';
-import { SummaryWord } from '../common/SummaryWord';
 import { WordCard } from '../common/WordCard';
+import { Mediator } from '../renderer/Mediator';
+import { SummaryWord } from '../renderer/SummaryWord';
 
 import Book from './Book';
 import BookController from './BookController';
