+++ src/renderer/SummaryWord.ts
