+++ src/renderer/components/CardRenderer.tsx
@@ -11,9 +11,9 @@ import React, { useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 
 import { LayoutCard, LayoutComponent } from '../../common/LayoutCard';
-import { Mediator } from '../../common/Mediator';
-import { SummaryWord } from '../../common/SummaryWord';
 import { WordCard } from '../../common/WordCard';
+import { Mediator } from '../Mediator';
+import { SummaryWord } from '../SummaryWord';
 import { pushSelectedWordAction } from '../actions/SelectedWordsActions';
 import { State } from '../states/State';
